Use `batch` priority when clearing intermediate tables
For configurations with many bigquery connectors, such as our debezium
data pipelines to bigquery that replicate 100s of tables, we'd like to
reduce the number of `interactive` queries running as to reduce the
likelyhood of hitting concurrent query limits.

Since the intermediate table clearing is done on a best effort basis
and since the code is already built to handle stale data in the table,
(including cases where the data is in the temp streaming buffer),
we can safely queue up deletes in `batch` mode.